[refactor] make LinkedInApiScopesEncoder package private
